initial commit
I initially developed this plugin for a website I was working on a
while ago, however the developer who came in after me chose to go with
a more mainstream plugin. I’ve forgotten about it until now, so I’m
just uploading it as is and maybe I’ll clean it up a bit if I get the
chance.